Refactor: ExpectationSetting->CompositeExpectation
Given that the public API of Expectation consists solely of expectation
setting methods, and that the matching, verifying and other misc methods
aren't part of the public API, it makes sense to equate the Expectation
protocol implicitly to the ExpectationSetting protocol. If we were to
segregate interfaces, we might wanna have other internal-facing
interfaces like ExpectationMatching, etc.

By naming this CompositeExpectation, we don't need (as much) to expose
it or to include it in/update the documentation because as far as
clients are concerned, they still get back an Expectation (technically,
something that quacks like an Expectation).