persist cardinality instance across updates
Fixes #473. Earlier, whenever you qualified an Expectation with a
cardinality, a new Cardinality instance would replace the old one.
The moving of invocations recording to Cardinality in 59454a8 then meant
that the invocations recorded prior to updating cardinality would get
discarded along with the Cardinality instance, effectively being
forgotten.

We now update the quantifiers of the same Cardinality instance for each
Expectation in order to persist the recording of invocations across
updates to cardinality.