Patch app.js to work with jquery v1.9.0
Ideally this would be fixed in yard. I've opened a PR [1] with the fix,
but haven't yet had any response at all - hence why I'm patching this.

Note that we now have to checkout both app.js & jquery.js after the docs
are generated in order to override the versions of these files generated
by yard.

[1]: https://github.com/lsegal/yard/pull/1351