Remove docs & tests from gem to reduce its size by over 50%
I was looking through https://rubygems.org/gems/mocha/versions and
noticed the .gem's file size increased from 94kb in v1.7.0 to 219kb in
v1.8.0 - an increase of 125kb.

The culprit seems to be the addition of generated docs to the repo in
4a85555831c55961a5100be18b0fe0b555b51945. I don't think there reasonable
value in including the generated docs in the compiled .gem, especially
for the byte size increase. Gem users still can locally still produce
documentation with rdoc/ri. I've also removed the test directory, since
they are not useful to gem users.

I'm using the syntax from the gemspec used in `bundle gem` [1].

[1]: https://github.com/rubygems/bundler/blob/v2.1.4/lib/bundler/templates/newgem/newgem.gemspec.tt#L23-L27