Temporarily ignore Ruby v1.8.7 build failures
There seems to be some kind of SSL certificate error [1] when installing
the bundler gem. This issue [2] looks as if it describes what I'm
seeing. It's not obvious that this is going to be fixed by Travis any
time soon, so given that I'm thinking of dropping support for Ruby
v1.8.7 and/or moving the CI builds somewhere else, I've decided to
ignore the failures for now.

[1]: https://travis-ci.org/github/freerange/mocha/builds/750442638
[2]: https://travis-ci.community/t/ruby-1-8-7-and-ree-builds-broken-by-ssl-certificate-failure/10866