Fix misleading exception in `has_entry`
If `has_entry` is called without any arguments, it will raise an
exception like the following:

```
ArgumentError: Too many arguments; use either a single argument (must be a Hash) or two arguments (a key and a value).
```

This is because we don't check if the `options.length` is zero so when
it is, the `else` branch executes and prints the above error. We can fix
this by simply adding another branch to cover the case when
`options.length` is zero.