Avoid breaking change in psych v4 in ruby v3.1
ruby 3.1 was shipped with psych v4 which included a breaking change [1]
that was mentioned in the ruby v3.1 release notes [2] (see the "Other
notable changes since 3.0" section).

This commit constrains psych to < v4. I haven't investigated, but I
suspect the code using psych is in rubocop, so we might be able to avoid
adding this constraint if we update to a more recent version of rubocop.

[1]: https://bugs.ruby-lang.org/issues/17866
[2]: https://www.ruby-lang.org/en/news/2021/12/25/ruby-3-1-0-released/