Rely on Object#respond_to? in MockTest
Although this is a unit test, I think it's a bit simpler to rely on the
built-in implementation of Object#respond_to? rather than redefining
that method in some of these tests.

Also the real implementation of Object#respond_to? accepts an optional
2nd argument. We could change the relevant tests to allow such an
argument in the redefined versions of the method in these tests.
However, this doesn't work in Ruby v1.8.6, because blocks can't have
default values for arguments. The new approach means that this is a
non-issue.