Extract handle_method_call from method_missing
Continues the refactoring started in #549 on the road to better keyword
args matching. This is part of a refactor sketched out in #544.

The key change is to pass arguments and the block directly instead of
re-splatting them. That way, we would only need to set ruby2_keywords at
the edges, instead of with each splat (for contrast, see the initial
spike in #534, where ruby2_keywords had to be set at multiple layers).

The extracting of handle_method_call is not strictly necessary, but it
makes it clearer that method_missing is just one of the possible entry
points for a method call.