Add strict keyword arg matching config option
This option doesn't currently do anything, but in a subsequent commit it
will determine whether to perform strict keyword argument comparision.

This option is turned off by default to enable gradual adoption, but may
be turned on by default in the future.

This option is only available in Ruby >= v2.7.