Separate linting from test tasks
Previously the "rubocop" task provided by the rubocop library was added
as a dependency of the "test" task. This commit introduces a new "lint"
task which invokes the "rubocop" taask and adds the "lint" task as a
dependency of the "default" task.

This means that running the "test" task now *only* runs the unit tests &
acceptance tests; not the linting or the performance tests.