Separate lint from test jobs in CircleCI
This makes it easier to see whether the linting has passed or failed
when looking at CI results, e.g. in the GitHub PR annotations.

There's no point in running the linting for every version of Ruby. For
now we're running it against Ruby v2.2 which is the Rubocop
TargetRubyVersion.

There's no need to set MOCHA_OPTIONS=debug for the "lint" job, because
we're not actually invoking any of the Mocha Ruby code.