Remove unused arg for HashMethods#mocha_inspect
This was introduced in this commit [1] apparently as a precursor to the
addition of ParameterMatchers::HasKeys in this commit [2]. However, the
latter only actually used the new optional `wrapped` argument when
calling ArrayMethods#mocha_inspect, and so I think the change to
HashMethods#mocha_inspect was unnecessary.

Since HashMethods#mocha_inspect is not a published part of the API, this
shouldn't require a major version bump.

Thanks to @wasabigeek for highlighting [3] this issue in
wasabigeek/mocha#2.

[1]: d8f44bcadabb67c47dc33eed8ca1e7010e083485
[2]: 18d81047fde4800d907382d2ddd2f3fb52671fc9
[3]: https://github.com/wasabigeek/mocha/pull/2/files#r989194611