Always wrap Hash inspection in braces
Previously in the description of an invocation or of a parameter
matcher, a single Hash argument was displayed without wrapping braces.

The reasoning behind this seems to be lost in the mists of time, but I
suspect it pre-dated keyword arguments and was an attempt at
reflecting the way you don't/didn't need to wrap some Hash arguments
with braces when invoking a method. However, even if it did make sense
at one point, it no longer does now that we have keyword arguments, so
it feels simpler to always wrap a Hash argument in braces.

I'm hoping this will also make it easier to see how adding strict
keyword matching will change the behaviour.

Co-authored-by: Nicholas Koh <nicholas.koh@shopify.com>