Improve inspection of positional & keyword Hashes
With the introduction of strict keyword matching, we're explicitly
distinguishing between positional and keyword Hashes. However, this
meant the descriptions of Hash arguments in invocations and parameter
matchers had become confusing. This attempts to make that less confusing
by always wrapping a Hash in braces unless we *know* its a keyword Hash,
in which case don't wrap it in braces.

This means that in earlier versions of Ruby where we can't distinguish
between a positional and keyword Hash, it will always be displayed
wrapped in braces.

Co-authored-by: Nicholas Koh <nicholas.koh@shopify.com>