Remove deprecated mocha/setup.rb mechanism
Previously this was the recommended way to integrate Mocha with a test
framework. However, this mechanism was deprecated in #399 in favour of
requiring a file specific to the relevant test framework, i.e.
mocha/test_unit.rb or mocha/minitest.rb. Test frameworks like RSpec
integrate by requiring mocha/api.rb.

This old mechanism was deprecated in v1.10.0 [1,2] and so is safe to
remove now.

[1]: 36adf880a0b9302fb5e770161c2e962e7e20fe71
[2]: 388f44d778c2239b52ffe5c52ace253419f91bd3