Remove redundant deprecation disabling in MockTest
I'm not sure these calls to DeprecationDisabler.disable_deprecations
were ever necessary. Currently the relevant deprecation warnings are
only every displayed if the reinstate_undocumented_behaviour_from_v1_9
configuration option is true which is the default. Perhaps this wasn't
always the case and these calls are a vestige of that...? Anyway, they
certainly aren't needed now.