Fail fast when mock receives invocations in another test
This converts the deprecation warning introduced in
d66c2b501685fe827a2c5e2236475e78d8082c11 into the raising of a
StubbingError as promised in the deprecation warning.

The deprecation warning was introduced in Mocha v1.11.2 which was
released 2 years ago and we're about to do a major version bump, so it's
safe to make this change.

See #440 & #442.