Improve keyword argument deprecation warning
By including the source location of the stub definition.

This change is a necessarily a bit pervasive and unfortunately makes the
code and particularly the tests even more complicated. However, I think
it will really help developers fix the deprecation warnings and
(hopefully) it will only be necessary until the next significant
release at which point we should be able to remove it!