Deprecate reinstate_undocumented_behaviour_from_v1_9
This configuration option will be removed in the future.

If you are explicitly setting it to true, then you need to fix any
deprecation warnings highlighting code which is relying on Mocha v1.9
behaviour.

If you are explicitly setting it to false, then you can remove the call
and rely on the new default value.