ruby2_keywords is runtime dep not development dep
And so should be included as such in the gemspec; not in the Gemfile.

I've specified the '>= 0.0.5' constraint, because v0.0.5 is the version
we've been testing with.

I've used a '>=' constraint rather than a twiddle-wakka ('~> 0.0.5'),
because that's equivalent to the combination of '>= 0.0.5' and '< 0.1.0'
which seems a bit too restrictive. At the moment there's no reason to
suppose later versions of ruby2_keywords won't work with Mocha and we
can always add a constraint later if it turns out there is a problem.

Closes #581.