Simplify storage of MOCHA_OPTIONS
This has only ever supported `debug`, `verbose` & `skip_integration`
options. However, it currently only supports `debug`.

Storing these options in a Hash was always a bit over the top and since
it was the only reason we needed to disabled the Style/EachWithObject
Rubocop rule, it seems more sensible to use an Array instead.

I don't think that the "uniqueness" of the options in the
comma-separated list was ever a motivation for storing them in a Hash
and so an Array (rather than a Set) seems good enough.