Pin JRuby to v9.3.9.0 in CircleCI builds
Currently 4 tests are failing in JRuby v9.4.0.0, e.g. this build [1].

This is due to a JRuby issue [2] which is being tracked in #591.
Although I believe the JRuby issue has been fixed it's not obvious to me
when it will be made available in the "jruby:latest" that the CircleCI
build was using.

It's quite distracting to see the CircleCI build fail all the time due
to this issue, so I've decided to pin the JRuby version at the last
working version. I'll need to try to remember to un-pin it again once
the fix is available.

[1]: https://app.circleci.com/pipelines/github/freerange/mocha/346/workflows/63497482-5d48-4e7c-8336-886ed9bdc071/jobs/4741
[2]: https://github.com/jruby/jruby/issues/7492