Pin yard version to v0.9.28 to avoid ArgumentError
It seems as if the recent v0.9.29 release of yard broke my custom
template code. I plan to report this to yard and to come up with a fix,
but in the meantime this fixes the CircleCI build. Without this change
we were seeing the following error when attempting to generate the docs:

    ArgumentError: No such template for default/fulldoc/html