Remove invalid CircleCI token from badge URL
I'm not sure why/when this token became invalid, but I suspect it was a
while back when there were some security issues.

Although the docs say that you don't need a token in the badge URL for
public repos, this commit [1] suggests otherwise. However, I just tried
it and it seems to work without, so I'm going to try that for now.

[1]: 607c5aab541e58d668a8b674cf5f3a1b08bfa379