Use minitest vs mini_test everywhere
As noted by @zenspider in this comment [1], the rename is over 10 years
old, so it seems about time we update it!

I've also renamed `test/mini_test_result.rb` ->
`test/minitest_result_pre_v5.rb` to better explain its purpose. I did
initially think about removing this, but we're still running pre v5
versions of Minitest in the CI builds for Ruby v2.0 & v2.1, so I'm going
to leave that for now and address it separately.

[1]: https://github.com/freerange/mocha/issues/614#issuecomment-1653124225