Rename StubbedMethod#hide_original_method -> #prepare
Originally this method used to rename the original method to move it out
the way and thus "hide" it. However, now we "hide" it by prepending a
module "in-front" of the original method. Since this change, I've found
this method name confusing and in any case I'm thinking about changing
the behaviour to *always* prepend a module even when there's no original
method. I'm also thinking about inlining this method into
`#define_new_method`, because I'm not convinced it's very useful on its
own.