Rename StubbedMethod#store_original_method_visibility -> #set_visibility
Originally we stored the original method visibility in order to restore
the original method as well as setting the visibility of the stubbed
method which we still do. Now that we only do the latter, I think it
makes less sense to reference the "original method" in the method name.
I plan to rename the relevant instance variable in a subsequent commit.