Prepend module even if method doesn't exist
Previously we were only prepending a module if the method existed,
because if the method did not exist, we could define the stubbed method
directly on the object being stubbed. However, by always defining
stubbed methods on a prepended module, we make the origins of the
stubbed method definition clearer as well as avoiding conditional logic
in `StubbedMethod#prepare`.