Convert StubbedMethod#stub_method_owner into attr_reader
Since we now always prepend a module on which to define the stubbed
method, the `stub_method_owner` instance variable is always set and we
no longer need to fallback to `#original_method_owner`.

There was some slight awkwardness in `InstanceMethodTest`, because a lot
of it is testing the implementation rather than the behaviour. Also I
had to delete one of the tests for `#remove_new_method`, which is no
longer relevant.