Improve unit tests for Hash#mocha_inspect
I can't work out what the origins of the
`test_should_keep_spacing_between_key_value` test name was and I don't
think it's very helpful, so I have renamed it.

I suspect the `test_should_use_mocha_inspect_on_each_item` test name was
copied from `ArrayInspectTest`. The idea of an item doesn't make so much
sense for a `Hash` and so I have renamed it to make things clearer.

I've also added a unit test for the behaviour for a keyword hash.
Although this behaviour is well-covered by acceptance tests, I think
that adding this is worthwhile.