Fix regression in ParameterMatchers#has_entry, etc
This regression was caused by the fix for nested parameter matching for
keyword arguments in #648.

The arguments passed in to `#has_entry` (and probably other matchers)
were inadvertently triggering the behaviour in
`ParameterMatchers::PositionalOrKeywordHash` when really that was only
intended to be used for arguments passed directly to a stubbed method.

In #648, this line [1], which is effectively the same as using
`ParameterMatchers::Equal`, was replaced [2] by use of
`ParameterMatchers::HasEntries`. While the latter was suitable for
arguments passed directly to a stubbed method, it was not suitable for
arguments passed to matcher methods like `#has_entry`.

Now the behaviour in `ParameterMatchers::PositionalOrKeywordHash` is
only triggered for arguments passed directly to a stubbed method. This
is achieved using a new `top_level` argument for the `#to_matcher`
method.

This `top_level` is only set to `true` in `ParametersMatcher#matchers`
which is deep in the call stack from `Mock#method_missing` as part of
the expectation matching process.

I'm not completely happy with the name of `top_level`, but I can't think
of anything better at the moment.

Fixes #654.

[1]: https://github.com/freerange/mocha/blob/5324009b296939c0a0da9d05c13b823a6dceb643/lib/mocha/parameter_matchers/positional_or_keyword_hash.rb#L16
[2]: https://github.com/freerange/mocha/blob/f94e250414335d822ec8d0a1331cb25cf90bf837/lib/mocha/parameter_matchers/positional_or_keyword_hash.rb#L17