Fix regression when stubbed method invoked with no args
In this commit [1] which was part of #660 and was released in v2.4.3,
the new code didn't take into account that the `parameter` local
variable in `HasEntries#matches?` can be `nil` if the stubbed method
expects a `Hash` or keyword arguments but is called with no arguments.

Previously this wasn't a problem, because the `nil` was handled by this
guard condition [2] in `HasEntry#matches?`. However, now we're calling
`#length` on `parameter` when `exact` is `true`, we need this new guard
condition in `HasEntries#matches?`.

[1]: https://github.com/freerange/mocha/commit/5e6a07b2710dac76e9346def561ca0d44765bf86
[2]: https://github.com/freerange/mocha/blob/e95fda25c41faaa4a7737b62a471544c5c2ddf2f/lib/mocha/parameter_matchers/has_entry.rb#L68