Add missing unit tests for HasEntries
The `HasEntries` matcher works when the actual argument is an instance
of `ActionController::Parameters` [1] rather than a `Hash` and we want
to retain that behaviour.

This is achieved by the `HasEntry` matcher requiring the `Hash`-like
object to respond to the `#keys` & `#[]` methods. This behaviour is now
tested by the tests added in this commit which make use of a new
`Hashlike` class which only implements `#keys` & `#[]`.

[1]: https://api.rubyonrails.org/classes/ActionController/Parameters.html