Fix exact match on ActionController::Parameters
This was broken in this previous commit [1] which was part of #660 and
which was released in v2.4.3.

Previously the only requirement we put on `Hash`-like arguments was that
they respond to `#keys` & `#[]` methods. However, the change in [1] we
introduced an implicit assumption that they also respond to a `#length`
method which is caused the problem described here [2].

However, `ActionController::Parameters` [3] is commonly used in Rails
apps and it does *not* respond to `#length`. So it seems better to add a
guard condition to ensure the object responds to `#keys` and use that to
determine the length.

Note that the `HasEntry`, `HasKey` and `HasKeys` matchers already work
this way and so this feels like it makes `HasEntries` more consistent.

[1]: https://github.com/freerange/mocha/commit/5e6a07b2710dac76e9346def561ca0d44765bf86
[2]: https://github.com/freerange/mocha/issues/662#issuecomment-2246105530
[3]: https://api.rubyonrails.org/classes/ActionController/Parameters.html