Fix frozen string literal warning in Mockery
Thanks to @radville for pointing this out in #669. I've tackled it
slightly differently by using an Array to build up the lines and then
joining them into a String at the end rather than calling `String#dup`
multiple times.

Previously I was seeing a lot of the following warnings when running
the tests under Ruby v3.4:

    lib/mocha/mockery.rb:118: warning: literal string will be frozen in the future