Fix rubocop version constraint for Ruby > v2.2.0
For some reason the `jaro_winkler` version constraint added in this
commit [2] was causing bundler to downgrade the `rubocop` version from
v0.58.2 to v0.56.0. Unfortunately this resulted in some `rubocop`
violations when running `bundle exec rake lint` which prevented me from
releasing. Adding this lower bound to the `rubocop` version seems to
solve the problem.

[1]: https://github.com/freerange/mocha/commit/5c7d14cb7d779bd22da3dbfe9d16ed1091ace7a1