Fix never expectation deprecation warning logic
In #681 I added logic to display a deprecation warning about a change I
planned to make in #679. However, it turns out the logic in both was
faulty as demonstrated in this example [1] where the deprecation
warning was incorrectly displayed for the 2nd call to `Foo.create_if`:

    class Foo
      def self.create_if(condition)
        new if condition
      end
    end

    test "it creates only if condition is true" do
      Foo.expects(:new).never
      Foo.create_if(false)

      Foo.expects(:new).once
      Foo.create_if(true)
    end

This commit adds a new acceptance test to cover an example where the
logic was incorrect and updates the logic in `Mock#handle_method_call`
to fix the logic so this new test passes and none of the existing tests
fail. Unfortunately the implementation is now rather complicated and
hard to follow, but I think it will do for now. I have a couple of ideas
for ways to simplify it, but I don't have time to work on that now.

I plan to update the equivalent logic in #679 to match the new logic
implemented here.

[1]: https://github.com/freerange/mocha/pull/679#issuecomment-2503272428