Ruby v3.4 stacktrace uses single-quote vs backtick
Since this change [1], lines in the stacktrace use an opening
single-quote instead of a backtick to delimit the method name, e.g.

    foo.rb:5:in 'Foo#bar': boom! (RuntimeError)

Instead of:

    foo.rb:5:in `Foo#bar': boom! (RuntimeError)

This commit makes a relevant assertion specific to the current Ruby
version.

This is somewhat similar to the fixes I put in place in #672, although
those were later considerably simplified in #680. Unfortunately I can't
see a way to achieve a similar simpification in this case.

[1]: https://bugs.ruby-lang.org/]issues/16495