Drop support for Ruby v2.1
This version has been EOL since 31 Mar 2017 [1], which is almost 8 years
ago.

`Mocha::Configuration#stubbing_method_on_nil` and `StubbingNilTest` are
now moot, because as of Ruby v2.2 `nil` is frozen.  The behaviour tested
in the latter is already tested in the more generic
`StubbingFrozenObjectTest`. See this commit [2] for details.

Closes #628.

[1]: https://www.ruby-lang.org/en/downloads/branches/
[2]: https://github.com/freerange/mocha/commit/a65ea1ed3dce43cbc2cc16b3660afd2cce3db33e