Indicate when matcher logic is defined by block
Previously the output from a test where such an exception was not
matched was a bit confusing, because no expected parameters were
displayed. While we can't easily display the logic from the block
itself, we can at least highlight that such logic does exist.

While this will change the content of some test failure messages, I have
previously specified that is not part of the documented API. [1]

[1]: https://github.com/freerange/mocha/issues/577