Allow some core Ruby methods to be overridden
These core Ruby method signatures violate
the `Style/OptionalBooleanParameter` cop and a bunch of tests
legitimately override them.

The one exception is `Mocha::ClassMethods::AnyInstance#respond_to?`
which is in library code. I've opened an issue (#713) to investigate
whether that class should define `#respond_to_missing?` instead.

Ideally I would've configured the extra `AllowedMethods` to only apply
to the tests, but I couldn't find a simple way to do that with the
rubocop configuration. It might be worth extracting a separate rubocop
configuration for the tests...?