Use built-in Ruby debug option vs custom module
The debug mode has become less important in recent years since we
stopped monkey-patching the test libraries.

The only downside I can see for this is that we'll see *all* Ruby debug
output when this option is enabled which might obscure Mocha's debug
output. However, I think we'll cross that bridge if/when we come to it.

I did consider setting `RUBYOPT=-d` in the `environment` entry in
`.circleci/config.yml` to replace `MOCHA_OPTIONS=debug`. However, this
would've cluttered up the CI build output unnecessarily.

Closes #714.