Default strict kwarg matching to true in Ruby v3
This changes the value of
`Configuration#strict_keyword_argument_matching?` depending on the Ruby
version. The default value will continue to be `false` in Ruby v2.7, but
it will now be `true` in Ruby >= v3.0. This configuration option is not
supported in Ruby < v2.7.

The whole purpose of this configuration option was to allow gradual
adaptation to strict keyword argument, i.e. playing a similar role as
Ruby v2.7 which was designed to allow developers to prepare for the
changes to keyword arguments in Ruby v3.0.

Previously the Mocha docs alluded to this by saying things like:

    This is because {#strict_keyword_argument_matching=} will default
    to +true+ in the future.

When strict matching is disabled, deprecation warnings are already
displayed when a test would fail if strict matching was enabled. So
there's no need to add a specific deprecation warning before changing
the configuration default, because the user could simply manually
disable strict matching to prevent their tests from failing.

Closes #697.